# src/lib/api/openapi-spec.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/me":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/domains":{"post":{"operationId":"DomainController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDomainDto"}}}},"responses":{"201":{"description":""}},"tags":["Domain"]},"get":{"operationId":"DomainController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Domain"]}},"/domains/{id}":{"get":{"operationId":"DomainController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Domain"]},"delete":{"operationId":"DomainController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Domain"]}},"/databases":{"post":{"operationId":"DatabaseController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDatabaseDto"}}}},"responses":{"201":{"description":""}},"tags":["Database"]},"get":{"operationId":"DatabaseController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Database"]}},"/databases/{id}":{"get":{"operationId":"DatabaseController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Database"]},"delete":{"operationId":"DatabaseController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Database"]}},"/droplets":{"post":{"operationId":"DropletController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDropletDto"}}}},"responses":{"201":{"description":""}},"tags":["Droplet"]},"get":{"operationId":"DropletController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Droplet"]}},"/droplets/{id}":{"get":{"operationId":"DropletController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Droplet"]},"delete":{"operationId":"DropletController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Droplet"]}},"/droplets/{id}/reboot":{"post":{"operationId":"DropletController_reboot","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Droplet"]}},"/billing/invoices":{"get":{"operationId":"BillingController_getInvoices","parameters":[],"responses":{"200":{"description":""}},"tags":["Billing"]}},"/billing/checkout":{"post":{"operationId":"BillingController_createCheckout","parameters":[],"responses":{"201":{"description":""}},"tags":["Billing"]}},"/billing/webhook":{"post":{"operationId":"BillingController_handleWebhook","parameters":[{"name":"stripe-signature","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Billing"]}}},"info":{"title":"Cloud Nexus API","description":"API documentation for Cloud Nexus Platform","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"}},"required":["email","password","name"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateDomainDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"CreateDatabaseDto":{"type":"object","properties":{"name":{"type":"string"},"engine":{"type":"string","enum":["postgres","mysql","redis"]},"version":{"type":"string"}},"required":["name","engine","version"]},"CreateDropletDto":{"type":"object","properties":{"name":{"type":"string"},"region":{"type":"string"},"size":{"type":"string"},"image":{"type":"string"}},"required":["name","region","size"]}}}}